+++ .env.local
@@ -1,3 +1,4 @@
-NOTION_API_KEY=
-NOTION_DATABASE_ID=
-NOTION_PAGE_ID=
+NOTION_API_KEY=secret_8JhuGHAYK0kyJyLSrgkYoHUnrA5bzTBrKZR5NzMkmD9
+NOTION_DATABASE_ID=31cc2824f8184e4db7bbcef415b25fd8
+NOTION_DATABASE_ID_FAQS=d7966baa2c50469db834b38c10d705ab
+NOTION_PAGE_ID=3797d457d5c343ff9487a6fc5efdc7c3
