+++ components/WaitingList.js
@@ -1,6 +1,6 @@
 function WaitingList () {
     return (
-        <div className="bg-white py-16 sm:py-24">
+        <div className="bg-white py-16 sm:py-24" id="sign-up">
             <div className="relative sm:py-16">
             <div aria-hidden="true" className="hidden sm:block">
                 <div className="absolute inset-y-0 left-0 w-1/2 bg-gray-50 rounded-r-3xl" />
@@ -20,7 +20,7 @@ function WaitingList () {
                 <rect width={404} height={392} fill="url(#8228f071-bcee-4ec8-905a-2a059a2cc4fb)" />
                 </svg>
             </div>
-            <div className="mx-auto max-w-md px-4 sm:max-w-3xl sm:px-6 lg:max-w-7xl lg:px-8">
+            <div className="mx-auto max-w-md px-4 sm:max-w-3xl sm:px-6 lg:max-w-7xl lg:px-8" >
                 <div className="relative rounded-2xl px-6 py-10 bg-indigo-600 overflow-hidden shadow-xl sm:px-12 sm:py-20">
                 <div aria-hidden="true" className="absolute inset-0 -mt-72 sm:-mt-32 md:mt-0">
                     <svg
@@ -51,16 +51,38 @@ function WaitingList () {
                         We are working on our app and before we publically launch we will reach out to our waiting list.
                     </p>
                     </div>
-                    <form action="#" className="mt-12 sm:mx-auto sm:max-w-lg sm:flex">
-                    <div className="min-w-0 flex-1">
-                        <label htmlFor="cta-email" className="sr-only">
-                        Email address
+                    <form action="#" className="mt-12 sm:mx-auto sm:flex">
+                    <div className="min-w-0 mt-4 sm:mt-0 sm:ml-3 flex-1">
+                        <label htmlFor="cta-first-name" className="sr-only">
+                            First
                         </label>
                         <input
                         id="cta-email"
                         type="email"
                         className="block w-full border border-transparent rounded-md px-5 py-3 text-base text-gray-900 placeholder-gray-500 shadow-sm focus:outline-none focus:border-transparent focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600"
-                        placeholder="Enter your email"
+                        placeholder="First"
+                        />
+                    </div>
+                    <div className="min-w-0 mt-4 sm:mt-0 sm:ml-3 flex-1">
+                        <label htmlFor="cta-last-name" className="sr-only">
+                            Last
+                        </label>
+                        <input
+                        id="cta-email"
+                        type="email"
+                        className="block w-full border border-transparent rounded-md px-5 py-3 text-base text-gray-900 placeholder-gray-500 shadow-sm focus:outline-none focus:border-transparent focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600"
+                        placeholder="Last"
+                        />
+                    </div>
+                    <div className="min-w-0 mt-4 sm:mt-0 sm:ml-3 flex-1">
+                        <label htmlFor="cta-email" className="sr-only">
+                            Email address
+                        </label>
+                        <input
+                        id="cta-email"
+                        type="email"
+                        className="block w-full border border-transparent rounded-md px-5 py-3 text-base text-gray-900 placeholder-gray-500 shadow-sm focus:outline-none focus:border-transparent focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600"
+                        placeholder="Email"
                         />
                     </div>
                     <div className="mt-4 sm:mt-0 sm:ml-3">
