+++ package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "@headlessui/react": "^1.4.2",
     "@heroicons/react": "^1.0.5",
+    "@iconify/react": "^3.1.0",
     "@notionhq/client": "^0.4.9",
     "@tailwindcss/typography": "^0.4.1",
     "heroicons-react": "^1.4.1",
