--- pages/about-us.js
@@ -1,548 +1,0 @@
-import {
-  InboxIcon,
-  SparklesIcon,
-  CheckCircleIcon
-} from '@heroicons/react/outline'
-
-import Footer from '../components/Footer'
-import WaitingList from '../components/WaitingList'
-import Header from '../components/Header'
-
-const stats = [
-    { label: 'Founded', value: '2021' },
-    { label: 'Employees', value: '5' },
-    { label: 'Beta Users', value: '521' },
-    { label: 'Raised', value: '$25M' },
-  ]
-
-const includedFeatures = [
-    'No exit fees',
-    'Top up when you want ',
-    'Free consolidation service',
-    'Trusted fund manager',
-]
-
-
-export default function AboutUs() {
-  return (
-    <div className="main-container">
-        <div className="bg-white">
-            <Header/>
-            <main>
-                <div>
-                {/* Hero card */}
-                <div className="relative">
-                    <div className="absolute inset-x-0 bottom-0 h-1/2 bg-gray-100" />
-                    <div className="max-w-7xl mx-auto sm:px-6 lg:px-8">
-                    <div className="relative shadow-xl sm:rounded-2xl sm:overflow-hidden">
-                        <div className="absolute inset-0">
-                        <img
-                            className="h-full w-full object-cover"
-                            src="/trees-hero.jpg"
-                            alt="Trees"
-                        />
-                        <div className="absolute inset-0 bg-indigo-700 mix-blend-multiply" />
-                        </div>
-                        <div className="relative px-4 py-16 sm:px-6 sm:py-24 lg:py-32 lg:px-8">
-                        <h1 className="text-center text-4xl font-extrabold tracking-tight sm:text-5xl lg:text-6xl">
-                            <span className="block text-white">Invest in our</span>
-                            <span className="block text-indigo-200">Planet</span>
-                        </h1>
-                        <p className="mt-6 max-w-lg mx-auto text-center text-xl text-indigo-200 sm:max-w-3xl">
-                            Easily consolidate your pensions into one simpler and greener plan, with a trusted provider. Gain full transparency over your pension and environmental impact.
-                        </p>
-                        <div className="mt-10 max-w-sm mx-auto sm:max-w-none sm:flex sm:justify-center">
-                            <div className="space-y-4 sm:space-y-0 sm:mx-auto sm:inline-grid sm:grid-cols-1 sm:gap-5">
-                            <a
-                                href="#"
-                                className="flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-indigo-700 bg-white hover:bg-indigo-50 sm:px-8"
-                            >
-                                Join the waiting list
-                            </a>
-                            {/* <a
-                                href="#"
-                                className="flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-indigo-500 bg-opacity-60 hover:bg-opacity-70 sm:px-8"
-                            >
-                                Live demo
-                            </a> */}
-                            </div>
-                        </div>
-                        </div>
-                    </div>
-                    </div>
-                </div>
-
-                {/* Logo cloud */}
-                <div className="bg-gray-100">
-                    <div className="max-w-7xl mx-auto py-16 px-4 sm:px-6 lg:px-8">
-                    <p className="text-center text-sm font-semibold uppercase text-gray-500 tracking-wide">
-                        Partnered with
-                    </p>
-                    <div className="mt-6 grid grid-cols-2 gap-8 md:grid-cols-6 lg:grid-cols-5">
-                        <div className="col-span-1 flex justify-center md:col-span-2 lg:col-span-1">
-                        <img className="h-12" src="https://tailwindui.com/img/logos/tuple-logo-gray-400.svg" alt="Tuple" />
-                        </div>
-                        <div className="col-span-1 flex justify-center md:col-span-2 lg:col-span-1">
-                        <img className="h-12" src="https://tailwindui.com/img/logos/mirage-logo-gray-400.svg" alt="Mirage" />
-                        </div>
-                        <div className="col-span-1 flex justify-center md:col-span-2 lg:col-span-1">
-                        <img
-                            className="h-12"
-                            src="https://tailwindui.com/img/logos/statickit-logo-gray-400.svg"
-                            alt="StaticKit"
-                        />
-                        </div>
-                        <div className="col-span-1 flex justify-center md:col-span-2 md:col-start-2 lg:col-span-1">
-                        <img
-                            className="h-12"
-                            src="https://tailwindui.com/img/logos/transistor-logo-gray-400.svg"
-                            alt="Transistor"
-                        />
-                        </div>
-                        <div className="col-span-2 flex justify-center md:col-span-2 md:col-start-4 lg:col-span-1">
-                        <img
-                            className="h-12"
-                            src="https://tailwindui.com/img/logos/workcation-logo-gray-400.svg"
-                            alt="Workcation"
-                        />
-                        </div>
-                    </div>
-                    </div>
-                </div>
-                </div>
-
-                {/* More main page content here... */}
-            </main>
-        </div>
-        <div className="bg-gray-50 pt-12 sm:pt-16">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="max-w-4xl mx-auto text-center">
-          <h2 className="text-3xl font-extrabold text-gray-900 sm:text-4xl">
-            Impact Calculator
-          </h2>
-          <p className="mt-3 text-xl text-gray-500 sm:mt-4">
-            Research by Bailiie Gifford. Based upon move the total value of assets to our Better World plan.  
-          </p>
-        </div>
-      </div>
-    
-      <div className="mt-10 pb-12 bg-white sm:pb-16">
-        <div className="relative">
-          <div className="absolute inset-0 h-1/2 bg-gray-50" />
-          <div className="relative max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="max-w-4xl mx-auto">
-              <dl className="rounded-lg bg-white shadow-lg sm:grid sm:grid-cols-3">
-                <div className="flex flex-col border-b border-gray-100 p-6 text-center sm:border-0 sm:border-r">
-                  <dt className="order-2 mt-2 text-lg leading-6 font-medium text-gray-500">Litres of drinking water
-saved</dt>
-                  <dd className="order-1 text-5xl font-extrabold text-indigo-600">132,095</dd>
-                </div>
-                <div className="flex flex-col border-t border-b border-gray-100 p-6 text-center sm:border-0 sm:border-l sm:border-r">
-                  <dt className="order-2 mt-2 text-lg leading-6 font-medium text-gray-500">Metrics tons of CO2
-avoided</dt>
-                  <dd className="order-1 text-5xl font-extrabold text-indigo-600">12</dd>
-                </div>
-                <div className="flex flex-col border-t border-gray-100 p-6 text-center sm:border-0 sm:border-l">
-                  <dt className="order-2 mt-2 text-lg leading-6 font-medium text-gray-500">Trees planted</dt>
-                  <dd className="order-1 text-5xl font-extrabold text-indigo-600">10,340</dd>
-                </div>
-              </dl>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div className="relative bg-white pt-16 pb-32 overflow-hidden">
-      <div className="relative">
-        <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
-          <div className="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0">
-            <div>
-              <div>
-                <span className="h-12 w-12 rounded-md flex items-center justify-center bg-indigo-600">
-                  <InboxIcon className="h-6 w-6 text-white" aria-hidden="true" />
-                </span>
-              </div>
-              <div className="mt-6">
-                <h2 className="text-3xl font-extrabold tracking-tight text-gray-900">
-                    Easily consolidate
-                </h2>
-                <p className="mt-4 text-lg text-gray-500">
-                    Tell us about your pension history and we'll transfer everything over automatically.
-                </p>
-                <div className="mt-6">
-                  <a
-                    href="#"
-                    className="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-700"
-                  >
-                    See more
-                  </a>
-                </div>
-              </div>
-            </div>
-            {/* <div className="mt-8 border-t border-gray-200 pt-6">
-              <blockquote>
-                <div>
-                  <p className="text-base text-gray-500">
-                    &ldquo;Cras velit quis eros eget rhoncus lacus ultrices sed diam. Sit orci risus aenean curabitur
-                    donec aliquet. Mi venenatis in euismod ut.&rdquo;
-                  </p>
-                </div>
-                <footer className="mt-3">
-                  <div className="flex items-center space-x-3">
-                    <div className="flex-shrink-0">
-                      <img
-                        className="h-6 w-6 rounded-full"
-                        src="https://images.unsplash.com/photo-1509783236416-c9ad59bae472?ixlib=rb-=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=8&w=1024&h=1024&q=80"
-                        alt=""
-                      />
-                    </div>
-                    <div className="text-base font-medium text-gray-700">Marcia Hill, Digital Marketing Manager</div>
-                  </div>
-                </footer>
-              </blockquote>
-            </div> */}
-          </div>
-          <div className="mt-12 sm:mt-16 lg:mt-0">
-            <div className="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
-              <img
-                className="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:left-0 lg:h-full lg:w-auto lg:max-w-none"
-                src="https://tailwindui.com/img/component-images/inbox-app-screenshot-1.jpg"
-                alt="Inbox user interface"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="mt-24">
-        <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
-          <div className="px-4 max-w-xl mx-auto sm:px-6 lg:py-32 lg:max-w-none lg:mx-0 lg:px-0 lg:col-start-2">
-            <div>
-              <div>
-                <span className="h-12 w-12 rounded-md flex items-center justify-center bg-indigo-600">
-                  <SparklesIcon className="h-6 w-6 text-white" aria-hidden="true" />
-                </span>
-              </div>
-              <div className="mt-6">
-                <h2 className="text-3xl font-extrabold tracking-tight text-gray-900">
-                    Choose your green pension strategy
-                </h2>
-                <p className="mt-4 text-lg text-gray-500">
-                    Choose either a reccommended strategy for you or choose from a number of our sustainable pension plans. Green plans, without the comprimise.
-                </p>
-                <div className="mt-6">
-                  <a
-                    href="#"
-                    className="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-700"
-                  >
-                    See more
-                  </a>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="mt-12 sm:mt-16 lg:mt-0 lg:col-start-1">
-            <div className="pr-4 -ml-48 sm:pr-6 md:-ml-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
-              <img
-                className="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:right-0 lg:h-full lg:w-auto lg:max-w-none"
-                src="https://tailwindui.com/img/component-images/inbox-app-screenshot-2.jpg"
-                alt="Customer profile user interface"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div className="relative bg-white pt-16 pb-32 overflow-hidden">
-      <div className="relative">
-        <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
-          <div className="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0">
-            <div>
-              <div>
-                <span className="h-12 w-12 rounded-md flex items-center justify-center bg-indigo-600">
-                  <InboxIcon className="h-6 w-6 text-white" aria-hidden="true" />
-                </span>
-              </div>
-              <div className="mt-6">
-                <h2 className="text-3xl font-extrabold tracking-tight text-gray-900">
-                    Track your progress
-                </h2>
-                <p className="mt-4 text-lg text-gray-500">
-                    Keep on top of your pension progress and top up when you want
-                </p>
-                <div className="mt-6">
-                  <a
-                    href="#"
-                    className="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-700"
-                  >
-                    See more
-                  </a>
-                </div>
-              </div>
-            </div>
-            {/* <div className="mt-8 border-t border-gray-200 pt-6">
-              <blockquote>
-                <div>
-                  <p className="text-base text-gray-500">
-                    &ldquo;Cras velit quis eros eget rhoncus lacus ultrices sed diam. Sit orci risus aenean curabitur
-                    donec aliquet. Mi venenatis in euismod ut.&rdquo;
-                  </p>
-                </div>
-                <footer className="mt-3">
-                  <div className="flex items-center space-x-3">
-                    <div className="flex-shrink-0">
-                      <img
-                        className="h-6 w-6 rounded-full"
-                        src="https://images.unsplash.com/photo-1509783236416-c9ad59bae472?ixlib=rb-=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=8&w=1024&h=1024&q=80"
-                        alt=""
-                      />
-                    </div>
-                    <div className="text-base font-medium text-gray-700">Marcia Hill, Digital Marketing Manager</div>
-                  </div>
-                </footer>
-              </blockquote>
-            </div> */}
-          </div>
-          <div className="mt-12 sm:mt-16 lg:mt-0">
-            <div className="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
-              <img
-                className="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:left-0 lg:h-full lg:w-auto lg:max-w-none"
-                src="https://tailwindui.com/img/component-images/inbox-app-screenshot-1.jpg"
-                alt="Inbox user interface"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="mt-24">
-        <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
-          <div className="px-4 max-w-xl mx-auto sm:px-6 lg:py-32 lg:max-w-none lg:mx-0 lg:px-0 lg:col-start-2">
-            <div>
-              <div>
-                <span className="h-12 w-12 rounded-md flex items-center justify-center bg-indigo-600">
-                  <SparklesIcon className="h-6 w-6 text-white" aria-hidden="true" />
-                </span>
-              </div>
-              <div className="mt-6">
-                <h2 className="text-3xl font-extrabold tracking-tight text-gray-900">
-                    Stay updated
-                </h2>
-                <p className="mt-4 text-lg text-gray-500">
-                    Receive regular updates about the positive impact you are making
-                </p>
-                <div className="mt-6">
-                  <a
-                    href="#"
-                    className="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-700"
-                  >
-                    See more
-                  </a>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="mt-12 sm:mt-16 lg:mt-0 lg:col-start-1">
-            <div className="pr-4 -ml-48 sm:pr-6 md:-ml-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
-              <img
-                className="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:right-0 lg:h-full lg:w-auto lg:max-w-none"
-                src="https://tailwindui.com/img/component-images/inbox-app-screenshot-2.jpg"
-                alt="Customer profile user interface"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div className="bg-gray-100">
-      <div className="pt-12 sm:pt-16 lg:pt-20">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center">
-            <h2 className="text-3xl font-extrabold text-gray-900 sm:text-4xl lg:text-5xl">Simple no-tricks pricing</h2>
-            <p className="mt-4 text-xl text-gray-600">
-                Unlike other pension providers, we have no hidden fees or exit fees. 
-            </p>
-          </div>
-        </div>
-      </div>
-      <div className="mt-8 bg-white pb-16 sm:mt-12 sm:pb-20 lg:pb-28">
-        <div className="relative">
-          <div className="absolute inset-0 h-1/2 bg-gray-100" />
-          <div className="relative max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="max-w-lg mx-auto rounded-lg shadow-lg overflow-hidden lg:max-w-none lg:flex">
-              <div className="flex-1 bg-white px-6 py-8 lg:p-12">
-                <h3 className="text-2xl font-extrabold text-gray-900 sm:text-3xl">Better World Plan</h3>
-                <p className="mt-6 text-base text-gray-500">
-                  Lorem ipsum dolor sit amet consect etur adipisicing elit. Itaque amet indis perferendis blanditiis
-                  repellendus etur quidem assumenda.
-                </p>
-                <div className="mt-8">
-                  <div className="flex items-center">
-                    <h4 className="flex-shrink-0 pr-4 bg-white text-sm tracking-wider font-semibold uppercase text-indigo-600">
-                      What's included
-                    </h4>
-                    <div className="flex-1 border-t-2 border-gray-200" />
-                  </div>
-                  <ul role="list" className="mt-8 space-y-5 lg:space-y-0 lg:grid lg:grid-cols-2 lg:gap-x-8 lg:gap-y-5">
-                    {includedFeatures.map((feature) => (
-                      <li key={feature} className="flex items-start lg:col-span-1">
-                        <div className="flex-shrink-0">
-                          <CheckCircleIcon className="h-5 w-5 text-green-400" aria-hidden="true" />
-                        </div>
-                        <p className="ml-3 text-sm text-gray-700">{feature}</p>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-              <div className="py-8 px-6 text-center bg-gray-50 lg:flex-shrink-0 lg:flex lg:flex-col lg:justify-center lg:p-12">
-                <p className="text-lg leading-6 font-medium text-gray-900">Better World Plan</p>
-                <div className="mt-4 flex items-center justify-center text-5xl font-extrabold text-gray-900">
-                  <span>0.59</span>
-                  <span className="ml-3 text-xl font-medium text-gray-500">%</span>
-                </div>
-                <p className="mt-4 text-sm">
-                  <a className="font-medium text-gray-500">
-                    Annual fee
-                  </a>
-                </p>
-                <div className="mt-6">
-                  <div className="rounded-md shadow">
-                    <a
-                      href="#"
-                      className="flex items-center justify-center px-5 py-3 border border-transparent text-base font-medium rounded-md text-white bg-gray-800 hover:bg-gray-900"
-                    >
-                      Sign up
-                    </a>
-                  </div>
-                </div>
-                {/* <div className="mt-4 text-sm">
-                  <a href="#" className="font-medium text-gray-900">
-                    Get a free sample <span className="font-normal text-gray-500">(20MB)</span>
-                  </a>
-                </div> */}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div className="relative bg-white py-16 sm:py-24">
-      <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:gap-24 lg:items-start">
-        <div className="relative sm:py-16 lg:py-0">
-          <div aria-hidden="true" className="hidden sm:block lg:absolute lg:inset-y-0 lg:right-0 lg:w-screen">
-            <div className="absolute inset-y-0 right-1/2 w-full bg-gray-50 rounded-r-3xl lg:right-72" />
-            <svg
-              className="absolute top-8 left-1/2 -ml-3 lg:-right-8 lg:left-auto lg:top-12"
-              width={404}
-              height={392}
-              fill="none"
-              viewBox="0 0 404 392"
-            >
-              <defs>
-                <pattern
-                  id="02f20b47-fd69-4224-a62a-4c9de5c763f7"
-                  x={0}
-                  y={0}
-                  width={20}
-                  height={20}
-                  patternUnits="userSpaceOnUse"
-                >
-                  <rect x={0} y={0} width={4} height={4} className="text-gray-200" fill="currentColor" />
-                </pattern>
-              </defs>
-              <rect width={404} height={392} fill="url(#02f20b47-fd69-4224-a62a-4c9de5c763f7)" />
-            </svg>
-          </div>
-          <div className="relative mx-auto max-w-md px-4 sm:max-w-3xl sm:px-6 lg:px-0 lg:max-w-none lg:py-20">
-            {/* Testimonial card*/}
-            <div className="relative pt-64 pb-10 rounded-2xl shadow-xl overflow-hidden">
-              <img
-                className="absolute inset-0 h-full w-full object-cover"
-                src="https://images.unsplash.com/photo-1521510895919-46920266ddb3?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&fp-x=0.5&fp-y=0.6&fp-z=3&width=1440&height=1440&sat=-100"
-                alt=""
-              />
-              <div className="absolute inset-0 bg-indigo-500 mix-blend-multiply" />
-              <div className="absolute inset-0 bg-gradient-to-t from-indigo-600 via-indigo-600 opacity-90" />
-              <div className="relative px-8">
-                <div>
-                  <img
-                    className="h-12"
-                    src="https://tailwindui.com/img/logos/workcation.svg?color=white"
-                    alt="Workcation"
-                  />
-                </div>
-                <blockquote className="mt-8">
-                  <div className="relative text-lg font-medium text-white md:flex-grow">
-                    <svg
-                      className="absolute top-0 left-0 transform -translate-x-3 -translate-y-2 h-8 w-8 text-indigo-400"
-                      fill="currentColor"
-                      viewBox="0 0 32 32"
-                      aria-hidden="true"
-                    >
-                      <path d="M9.352 4C4.456 7.456 1 13.12 1 19.36c0 5.088 3.072 8.064 6.624 8.064 3.36 0 5.856-2.688 5.856-5.856 0-3.168-2.208-5.472-5.088-5.472-.576 0-1.344.096-1.536.192.48-3.264 3.552-7.104 6.624-9.024L9.352 4zm16.512 0c-4.8 3.456-8.256 9.12-8.256 15.36 0 5.088 3.072 8.064 6.624 8.064 3.264 0 5.856-2.688 5.856-5.856 0-3.168-2.304-5.472-5.184-5.472-.576 0-1.248.096-1.44.192.48-3.264 3.456-7.104 6.528-9.024L25.864 4z" />
-                    </svg>
-                    <p className="relative">
-                      Tincidunt integer commodo, cursus etiam aliquam neque, et. Consectetur pretium in volutpat, diam.
-                      Montes, magna cursus nulla feugiat dignissim id lobortis amet.
-                    </p>
-                  </div>
-
-                  <footer className="mt-4">
-                    <p className="text-base font-semibold text-indigo-200">Sarah Williams, CEO at Workcation</p>
-                  </footer>
-                </blockquote>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="relative mx-auto max-w-md px-4 sm:max-w-3xl sm:px-6 lg:px-0">
-          {/* Content area */}
-          <div className="pt-12 sm:pt-16 lg:pt-20">
-            <h2 className="text-3xl text-gray-900 font-extrabold tracking-tight sm:text-4xl">
-              On a mission to empower teams
-            </h2>
-            <div className="mt-6 text-gray-500 space-y-6">
-              <p className="text-lg">
-                Sagittis scelerisque nulla cursus in enim consectetur quam. Dictum urna sed consectetur neque tristique
-                pellentesque. Blandit amet, sed aenean erat arcu morbi. Cursus faucibus nunc nisl netus morbi vel
-                porttitor vitae ut. Amet vitae fames senectus vitae.
-              </p>
-              <p className="text-base leading-7">
-                Sollicitudin tristique eros erat odio sed vitae, consequat turpis elementum. Lorem nibh vel, eget
-                pretium arcu vitae. Eros eu viverra donec ut volutpat donec laoreet quam urna. Sollicitudin tristique
-                eros erat odio sed vitae, consequat turpis elementum. Lorem nibh vel, eget pretium arcu vitae. Eros eu
-                viverra donec ut volutpat donec laoreet quam urna.
-              </p>
-              <p className="text-base leading-7">
-                Rhoncus nisl, libero egestas diam fermentum dui. At quis tincidunt vel ultricies. Vulputate aliquet
-                velit faucibus semper. Pellentesque in venenatis vestibulum consectetur nibh id. In id ut tempus
-                egestas. Enim sit aliquam nec, a. Morbi enim fermentum lacus in. Viverra.
-              </p>
-            </div>
-          </div>
-
-          {/* Stats section */}
-          <div className="mt-10">
-            <dl className="grid grid-cols-2 gap-x-4 gap-y-8">
-              {stats.map((stat) => (
-                <div key={stat.label} className="border-t-2 border-gray-100 pt-6">
-                  <dt className="text-base font-medium text-gray-500">{stat.label}</dt>
-                  <dd className="text-3xl font-extrabold tracking-tight text-gray-900">{stat.value}</dd>
-                </div>
-              ))}
-            </dl>
-            <div className="mt-10">
-              <a href="#" className="text-base font-medium text-indigo-600">
-                {' '}
-                Learn more about how we're changing the world <span aria-hidden="true">&rarr;</span>{' '}
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <WaitingList/>
-    <Footer/>
-    </div>
-  )
-}
